+++ GrailDiary/BookImporter/BookImporter.swift
@@ -15,11 +15,6 @@ final class BookImporter {
   let database: NoteDatabase
   let apiKey: String?
 
-  struct BookAndImage {
-    var book: AugmentedBook
-    var image: TypedData?
-  }
-
   private var booksAndImages = [BookAndImage]()
   private let imageCache = ImageCache()
 
@@ -68,27 +63,7 @@ final class BookImporter {
 
   private func saveBooksAndImages(hashtags: String) {
     do {
-      try database.bulkUpdate(updateBlock: { db, updateIdentifier in
-        for bookAndImage in booksAndImages {
-          let identifier = UUID().uuidString
-          var note = Note(bookAndImage, hashtags: hashtags)
-          let noteTimestamp = bookAndImage.book.dateAdded ?? Date()
-          note.metadata.creationTimestamp = noteTimestamp
-          note.metadata.modifiedTimestamp = noteTimestamp
-          note.metadata.book = bookAndImage.book
-          try note.save(identifier: identifier, updateKey: updateIdentifier, to: db)
-          if let typedData = bookAndImage.image {
-            let binaryRecord = BinaryContentRecord(
-              blob: typedData.data,
-              noteId: identifier,
-              key: Note.coverImageKey,
-              role: "embeddedImage",
-              mimeType: typedData.type.preferredMIMEType ?? "application/octet-stream"
-            )
-            try binaryRecord.save(db)
-          }
-        }
-      })
+      try database.bulkImportBooks(booksAndImages, hashtags: hashtags)
     } catch {
       Logger.shared.error("Error importing books: \(error)")
     }
@@ -104,32 +79,10 @@ final class BookImporter {
   private func openLibraryCoverPublisher(for bookInfo: AugmentedBook, isbn: String) -> AnyPublisher<BookAndImage, Never> {
     return OpenLibrary.coverImagePublisher(isbn: isbn)
       .map { BookAndImage(book: bookInfo, image: $0) }
-      .catch { error -> Just<BookImporter.BookAndImage> in
+      .catch { error -> Just<BookAndImage> in
         Logger.shared.error("Error getting image for book \(isbn): \(error)")
         return Just(BookAndImage(book: bookInfo, image: nil))
       }
       .eraseToAnyPublisher()
-  }
-}
-
-extension Note {
-  init(_ bookAndImage: BookImporter.BookAndImage, hashtags: String) {
-    let book = bookAndImage.book
-    var markdown = ""
-    if let review = book.review {
-      markdown += "\(review)\n\n"
-    }
-    if let rating = book.rating, rating > 0 {
-      markdown += "#rating/" + String(repeating: "⭐️", count: rating) + " "
-    }
-    if !hashtags.trimmingCharacters(in: .whitespaces).isEmpty {
-      markdown += "\(hashtags)\n\n"
-    }
-    if let tags = book.tags {
-      for tag in tags {
-        markdown += "\(tag)\n"
-      }
-    }
-    self.init(markdown: markdown)
   }
 }
